feat: return the same error from decrypt_protected on invalid MDC
Previously checksum::sha1() called ensure_eq! and the resulting error
propagated upwards.  As it can end up in the logs or UI, it is better
to make it more generic to avoid accidental information leaks.

Additionally, compute SHA1 in advance, before performing the checks.
In the future, these checks can be made constant time, even though it
is out of scope of rPGP threat model for now.