Implement DSA signing as well
This also adds more tests, using the test vectors from RFC6979.

This adds a dependency on the hmac-drbg crate. The ecdsa crate also
needs this functionality, and in a comment they mention wanting to
switch to this crate after a pull request (which has already been
merged) is merged.

The current implementation in the ecdsa crate is not usable, since it
assumes the hash output size will always be larger than the desired
output size, and this is not the case for small hash functions with big
DSA keys. The hmac-drbg crate does implement this correctly. Besides
that the implementations seem interchangeable.