feat: add a method to construct a PublicKey from parameters
* Add a method to construct an PublicKey from parameters.
Currently the only way to create a PublicKey is to have a
pre-existing packet or to generate a new SecretKey and get it
from there. This change adds a method that allows PublicKey
objects to be synthesised from data from other sources, for
example external key material.

* Convert the from_params constructor to a more general one.
This permits the underlying key and packet version to be provided
as well.

* Reimplement ::from_slice in terms of ::new.
To achieve this we add algorithm as an argument of ::new and
change the return type to be Result<Self> so we can move the
version/algorithm check in to ::new.