fix!: generalize issuer subpacket checks
A signature may contain multiple issuer and issuer fingerprint subpackets, all of them should be considered when checking if a key could be the issuer of a signature.

Note: we drop the issuer subpacket checks in verify_key_binding_internal(): Key bindings should always be self-signatures. Gating on issuer subpackets is not useful, and may even cause wrong rejections of some back-signatures.