feat: obfuscate symmetric key length with ecdh padding
Padding in https://www.rfc-editor.org/rfc/rfc6637 is described as being based on PKCS5. However, RFC 6337 allows padding that exceeds the size of one block, with the goal of obfuscating the size of the symmetric key material.

This commit changes the default size of padded payloads to 40 bytes, which is the regular size of a padded payload that contains a AES256 key.

fixes #292