Freshen repo's freshness scent
This PR is purely to freshness the scent conveyed by the readme. If merged, this PR removes the false stale scent therein.

**Previously**: the readme displays "Last updated: October 2019" in Heading 2 font, above the fold as the visitor starts reading.

This conveyed that this repo might be abandonware, which is the impression I conjured on my first TWO visits to the repo home page.

Then repeatedly seeing links to this repo on socials I thought, why are all these people linking to abandonware?  It's only on my THIRD time that I realized, it's just the Heading 2 that's sending the stale signal.

This PR updates the last time this particular section was updated, February 8 2024.

A pertinent question may be: why display a date in the H2 heading at all?

This PR freshens the date so at least the stale scent is removed.